Fix host folder binding for Windows Docker Desktop
Unix-like paths as printed by `pwd` in Git Bash are incompatible with
Docker Desktop on Windows; they cause host binding to fail and create
empty dummy directories like 'cs2110;C'. Instead, we use `pwd -W` to
format it as a Windows 'C:'-style path, and fall back on `pwd` if that's
unavailable.