Add fixup image and fix manpages
There are several cases where we want to patch an image in the middle of
the semester but don't want the students that don't know how to use
docker to have to spend a bunch of space when pulling the new image when
our layers don't cache.

This adds a `fixup` image on top of the regular latest image. Here, we
fix the issue of manpages being removed in minimal Ubuntu. The
`/usr/bin/man` executable is also diverted to a warning shell script, so
we undivert it to replace it with the correct binary.