Fix stopping/removing existing containers
Use arrays to fix the previous script's invalid code for
stopping/removing existing containers, which would fail with an error if
there were multiple containers since arguments aren't split on newlines
the way the code was expecting. This splits into an array instead and
expands the array into multiple arguments when necessary. Using arrays
is even more bash-reliant, but the script already was not
POSIX-compliant anyway.

Don't use unquoted variables in bash, folks!